+++ lua/tst/expert.lua
@@ -14,9 +14,8 @@ pico.check("expert-02")
 pico.output.present()
 pico.check("expert-03")
 
-local a = pico.get.ticks();
-pico.input.delay(100);
-local b = pico.get.ticks();
-assert(b>=a+100 and b<=a+110);  -- 110 b/c of valgrind
+-- test: delay returns dt
+local dt = pico.input.delay(100)
+assert(dt>=100 and dt<=110)  -- 110 b/c of valgrind
 
 pico.init(false)