+++ tst/quit.c
@@ -7,8 +7,7 @@ int main (void) {
     pico_quit();
 
     Pico_Event e;
-    int has = pico_input_event_timeout(&e, PICO_EVENT_QUIT, 0);
-    assert(has);
+    pico_input_event_timeout(&e, PICO_EVENT_QUIT, 0);
     assert(e.type == PICO_EVENT_QUIT);
 
     pico_init(0);